+++ graphkernel/hash_graph_kernel_parallel.py
@@ -40,10 +40,10 @@ def hash_graph_kernel_parallel(graph_db, base_kernel, kernel_parameters, hashing
     g = graph_db[0]
     v = list(graph_db[0].vertices())[0]
     dim_attributes = len(g.vp.na[v])
-    colors_0 = np.zeros([num_vertices, dim_attributes],dtype=np.float32)
+    colors_0 = np.zeros([num_vertices, dim_attributes],dtype=np.float64)
     offset = 0
 
-    gram_matrix = sparse.lil_matrix((n,n),dtype=np.float32).tocsr()
+    gram_matrix = sparse.lil_matrix((n,n),dtype=np.float64).tocsr()
 
     # Get attributes from all graph instances
     graph_indices = []
